==> Python code for chatGpt/01 NSE BSE Delivery Data.ipynb <==
from platform import python_version

print(python_version())

### This program gives NSE BSE delivery for any stock based on NSE code or BSE Code
#

master_file_path = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Python Code from ChatGPT\\Master File\\CompanyMaster.csv"
directory = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Data Files\\"  # Replace with the path to your directory


column_order = ['NSE_Code','BSE_Code','Date','NSE_Delivery','BSE_Delivery']


import os
import csv
import pandas as pd



def search_nse_code(directory, nse_code, no_days):
# Create an empty DataFrame to store the results
    nse_df = pd.DataFrame(columns=['NSE_Code', 'Date','NSE_Delivery'])
    dely_segment = ['EQ','BE','SM','ST']

    file_names = sorted([file for file in os.listdir(directory) if file.endswith('.csv') and file.startswith('NSE Delivery')], reverse=True)[:no_days]


    for file_name in file_names:
        file_path = os.path.join(directory, file_name)
        date = file_name[13:19]
            
        with open(file_path, 'r') as file:
            reader = csv.reader(file)
            found_match = False  # Flag variable to track the match
            for i, line in enumerate(reader, start=1):
                if i >= 5:  # Start reading from line 5
                    line_str = ','.join(line)
                    data = line_str.split(',')
                    if data[2] == nse_code and data[3] in dely_segment:
                        dely = data[5]
                        nse_df = nse_df.append({'NSE_Code': nse_code, 'Date': date, 'NSE_Delivery':dely}, ignore_index=True)
                        found_match = True
                        break
            if found_match:
                continue
            

    return nse_df

def search_bse_code(directory, bse_code, no_days):
    # Create an empty DataFrame to store the results
    bse_df = pd.DataFrame(columns=['BSE_Code', 'Date','BSE_Delivery'])
    dely_segment = ['EQ','BE','SM','ST']

    file_names = sorted([file for file in os.listdir(directory) if file.endswith('.txt') and file.startswith('BSE Gross Deliverable ')], reverse=True)[:no_days]


    for file_name in file_names:
        file_path = os.path.join(directory, file_name)
        date = file_name[22:28] 
    
        with open(file_path, 'r') as file:
            reader = csv.reader(file)
            found_match = False  # Flag variable to track the match
            for i, line in enumerate(reader, start=1):
                if i >= 2:  # Start reading from line 5
                    line_str = ','.join(line)
                    data = line_str.split('|')
                    if data[1] == bse_code:
                        dely = data[2]
                        bse_df = bse_df.append({'BSE_Code': bse_code, 'Date': date, 'BSE_Delivery':dely}, ignore_index=True)
                        found_match = True
                        break
            if found_match:
                continue
            
    return bse_df
                

def find_corresponding_codes(master_file_path, user_input):
    nse_code = ""
    bse_code = ""

    user_input = user_input.lower()  # Convert user input to lowercase for comparison

    with open(master_file_path, 'r') as file:
        reader = csv.reader(file)
        for row in reader:
            if row[1].lower() == user_input:  # NSE code match (case-insensitive)
                nse_code = row[1]
                bse_code = str(row[2])
                break
            elif row[2].lower() == user_input:  # BSE code match (case-insensitive)
                nse_code = row[1]
                bse_code = str(row[2])
                break

    return nse_code, bse_code

def format_value(value):
    return '{:,.0f}'.format(value)

def format_decimal(x):
    return '{:,.2f}'.format(x)




def search_nse_price(directory, nse_code, no_days):
# Create an empty DataFrame to store the results
    nse_df = pd.DataFrame(columns=['NSE_Code', 'Date','Price','Traded_Qty'])
    dely_segment = ['EQ','BE','SM','ST']

    file_names = sorted([file for file in os.listdir(directory) if file.lower().endswith(('CSV','.csv')) and file.startswith('NSE EQ')], reverse=True)[:no_days]


    for file_name in file_names:
        file_path = os.path.join(directory, file_name)
        date = file_name[7:13]
            
        with open(file_path, 'r') as file:
            reader = csv.reader(file)
            found_match = False  # Flag variable to track the match
            for row in reader:
                data_list = []
                data_list.append(row)
                    
                if data_list[0][0] == nse_code and data_list[0][1] in dely_segment:
                        price = data_list[0][5]
                        traded_qty = data_list[0][8]
                        nse_df = nse_df.append({'NSE_Code': nse_code, 'Date': date, 'Price':price, 
                                                'Traded_Qty': traded_qty}, ignore_index=True)
                        found_match = True
                        break
            if found_match:
                continue
            

    return nse_df

def search_bse_price(directory, bse_code, no_days):
# Create an empty DataFrame to store the results
    bse_df = pd.DataFrame(columns=['BSE_Code', 'Date','Price'])
    dely_segment = ['EQ','BE','SM','ST']

    file_names = sorted([file for file in os.listdir(directory) if file.lower().endswith(('CSV','csv')) and file.startswith('EQ ISIN')], reverse=True)[:no_days]


    for file_name in file_names:
        file_path = os.path.join(directory, file_name)
        date = file_name[8:14]
            
        with open(file_path, 'r') as file:
            reader = csv.reader(file)
            found_match = False  # Flag variable to track the match
            for row in reader:
                data_list = []
                data_list.append(row)
                    
                if data_list[0][0] == bse_code :
                        price = data_list[0][7]
                        bse_df = bse_df.append({'BSE_Code': bse_code, 'Date': date, 'Price':price}, ignore_index=True)
                        found_match = True
                        break
            if found_match:
                continue
            

    return bse_df





def multiple_scrips(scrip_list,no_days):
    
    
    for i in scrip_list:
        nse_code, bse_code = find_corresponding_codes(master_file_path, i)
    
    #### search for price info in nse/bse files 

        nse_price = search_nse_price(directory,nse_code,no_days)
        nse_price['Price'] = nse_price['Price'].astype(float)
    
        nse_result = search_nse_code(directory, nse_code,no_days)
        bse_result = search_bse_code(directory, bse_code,no_days)
    
    ######

        if nse_result.empty:
            if nse_code == "":
                print('stock not in nse')
            else:
                print('stock in be or st, merging price_info')
                nse_merge = pd.merge(nse_price, nse_result, on='Date', how = 'left')
                data = {'NSE_Code': nse_merge['NSE_Code_x'],
                        'Date': nse_merge['Date'],
                        'NSE_Delivery': nse_merge['Traded_Qty']
                        }
                nse_result = pd.DataFrame(data)
    
        if nse_result.empty:
            merged_df = pd.merge(bse_result, nse_result, on='Date', how='left')
            merged_df['NSE_Delivery'] = merged_df['NSE_Delivery'].fillna(0)
            merged_df['NSE_Code'] = merged_df['NSE_Code'].fillna("")
# Check if bse_result is empty
        elif bse_result.empty:
            merged_df = pd.merge(nse_result, bse_result, on='Date', how='left')
            merged_df['BSE_Delivery'] = merged_df['BSE_Delivery'].fillna(0)
            merged_df['BSE_Code'] = merged_df['BSE_Code'].fillna("")
# Neither nse_result nor bse_result is empty
        else:
            merged_df = pd.merge(nse_result, bse_result, on='Date', how='inner')
        
        merged_df = merged_df[column_order]

        merged_df['NSE_Delivery'] = merged_df['NSE_Delivery'].astype(int)
        merged_df['BSE_Delivery'] = merged_df['BSE_Delivery'].astype(int)
    
        merged_df['Total_Delivery'] = merged_df['NSE_Delivery'] + merged_df['BSE_Delivery']
    
        nse_price = search_nse_price(directory,nse_code,no_days)
        nse_price['Price'] = nse_price['Price'].astype(float)

        if nse_price.empty:
            bse_price = search_bse_price(directory,bse_code,no_days)
            bse_price['Price'] = bse_price['Price'].astype(float)
            final_df = pd.merge(merged_df, bse_price, on=['Date','BSE_Code'], how='inner')
        else:
            final_df = pd.merge(merged_df, nse_price, on=['Date','NSE_Code'], how = 'inner')
            final_df.drop(columns='Traded_Qty', inplace=True)
    
    
##format the df for final printing/display
        formatted_df = final_df.style.format({'NSE_Delivery': format_value, 'BSE_Delivery': format_value, 
                                          'Total_Delivery': format_value, 'Price' : format_decimal}) 


    #formatted_df
    # Append the formatted DataFrame to the list
    #formatted_dfs.append(formatted_df)
    
    # Print the formatted DataFrame
    #print(f"Formatted DataFrame for {i}:")
    #print(formatted_df)
    #print("\n")  # Add a new line for separation
    
        print(f"Formatted DataFrame for {i}:")
        display(formatted_df)
        print("\n")  # Add a new line for separation

    




nse_dely_file = sorted([file for file in os.listdir(directory) if file.endswith('.csv') and file.startswith('NSE Delivery')], reverse=True)[:1]
def extract_date(file_name,start):
    
    for i in file_name:
        end = start + 6
        dely_date = i[start:end]
        
    return dely_date

nse_dely_date = extract_date(nse_dely_file,13)
print(nse_dely_date)

##getting user input
default_value = 15
user_input = input("Enter NSE code or BSE code: ")
no_days = input(f"No of days for which data required:(default:{default_value})")

while not no_days.isdigit() or int(no_days) <= 0:
    print("Invalid input. Please enter a positive numeric value.")
    no_days = input("Enter a positive numeric value for no of days: ")

    #"Enter a number (default: {default_value}): "

no_days = int(no_days)


#### finding corresponding codes of nse and bse

nse_code, bse_code = find_corresponding_codes(master_file_path, user_input)

if nse_code and bse_code:
    print("Corresponding codes found:")
    print(f"NSE code: {nse_code}")
    print(f"BSE code: {bse_code}")
else:
    print("Code not found in the master file.")
    print(f"NSE code: {nse_code}")
    print(f"BSE code: {bse_code}")

    
#### search for price info in nse/bse files 

nse_price = search_nse_price(directory,nse_code,no_days)
nse_price['Price'] = nse_price['Price'].astype(float)


### finding nse and bse delivery information
nse_result = search_nse_code(directory, nse_code,no_days)
bse_result = search_bse_code(directory, bse_code,no_days)




######

if nse_result.empty:
    if nse_code == "":
        print('stock not in nse')
    
    else:
        print('stock in be or st, merging price_info')
        nse_merge = pd.merge(nse_price, nse_result, on='Date', how = 'left')
        data = {'NSE_Code': nse_merge['NSE_Code_x'],
                'Date': nse_merge['Date'],
                'NSE_Delivery': nse_merge['Traded_Qty']
                }
        nse_result = pd.DataFrame(data)


        
def calculate_delyqty(row):
    if pd.isna(row['NSE_Delivery']):
        return row['Traded_Qty']
    else:
        return row['NSE_Delivery']



        
#### this code checks if latest date is available. if not then it gets it. 
##### some data could be missing if stock has moved from be to eq-- need to fix that

if not nse_result.empty:
    # Find the maximum date string
    max_date_str = nse_result['Date'].max()

    if not pd.isna(max_date_str) and max_date_str != nse_dely_date:
        print('Performing the else loop based on the condition')
        nse_merge = pd.merge(nse_price, nse_result, on='Date', how = 'left')
        
        nse_merge['delyqty'] = nse_merge.apply(calculate_delyqty, axis=1)
    
        data = {'NSE_Code': nse_merge['NSE_Code_x'],
                'Date': nse_merge['Date'],
                'NSE_Delivery': nse_merge['delyqty']
                }
        nse_result_1 = pd.DataFrame(data)
        #nse_result_2 = nse_result
        #nse_result_merged = pd.concat([nse_result_1,nse_result_2], ignore_index=True)
        # Sort nse_result DataFrame by 'Date' column in descending order
        #nse_result = nse_result_merged.sort_values(by='Date', ascending=False)
        nse_result = nse_result_1

        # Reset the index of the sorted DataFrame
        nse_result.reset_index(drop=True, inplace=True)
        

#### 
    

# Perform inner join on 'Date' field based on delivery info found
if nse_result.empty:
    merged_df = pd.merge(bse_result, nse_result, on='Date', how='left')
    merged_df['NSE_Delivery'] = merged_df['NSE_Delivery'].fillna(0)
    merged_df['NSE_Code'] = merged_df['NSE_Code'].fillna("")
# Check if bse_result is empty
elif bse_result.empty:
    merged_df = pd.merge(nse_result, bse_result, on='Date', how='left')
    merged_df['BSE_Delivery'] = merged_df['BSE_Delivery'].fillna(0)
    merged_df['BSE_Code'] = merged_df['BSE_Code'].fillna("")
# Neither nse_result nor bse_result is empty
else:
    merged_df = pd.merge(nse_result, bse_result, on='Date', how='inner')

    
#### changing column order of merged df (column order defined in cell 1) and making delivery value as int
#### adding total delivery column

merged_df = merged_df[column_order]

merged_df['NSE_Delivery'] = merged_df['NSE_Delivery'].astype(int)
merged_df['BSE_Delivery'] = merged_df['BSE_Delivery'].astype(int)

merged_df['Total_Delivery'] = merged_df['NSE_Delivery'] + merged_df['BSE_Delivery']







################################
#### search for price info in nse/bse files and merge with delivery data

nse_price = search_nse_price(directory,nse_code,no_days)
nse_price['Price'] = nse_price['Price'].astype(float)

if nse_price.empty:
    bse_price = search_bse_price(directory,bse_code,no_days)
    bse_price['Price'] = bse_price['Price'].astype(float)
    final_df = pd.merge(merged_df, bse_price, on=['Date','BSE_Code'], how='inner')
else:
    final_df = pd.merge(merged_df, nse_price, on=['Date','NSE_Code'], how = 'inner')
    final_df.drop(columns='Traded_Qty', inplace=True)

    
    
##format the df for final printing/display
formatted_df = final_df.style.format({'NSE_Delivery': format_value, 'BSE_Delivery': format_value, 
                                      'Total_Delivery': format_value, 'Price' : format_decimal}) 


formatted_df
