==> Python code for chatGpt/50 NSE BSE Averages.ipynb <==
#### update companymaster file - check ipo listing from chittorgarh 

### run report designer in spider and save xls for bse, nse and sme

#### get marketcap from screener and save in marketcap from screener csv

#### run 57 Tradingview WL for percent change and new highs - this reads the file from tradingview
#### for 3 percent changes or more and new highs and makes a 0_tradingview_wl_yymmdd

#### run 50 nse bse averages - this takes nse/bse/sme averages and creates a combined average file (input file)

### run 51 combined nse bse files --- combines nse and bse files for a specific date - this combined file is used for 
#### report generation  -- this generates top 50 by value/volume for non-fno, fno and sme
#### this also creates non_fno_wl_yymmdd
#### this creates sme yymmdd 

#### run 56 delivery comparision over 2 days - this generates scrips where dely volumes are 3x of previous day
#### this generates 0_volume_3x_wl_yymmdd

#### run 53 report for multiple scrips
#### this takes 0_tradingview_wl_yymmdd (created in 57) and generate 7-day delivery report
#### this takes 0_volume_3x_wl_yymmdd (created in 56) and generate 7-day delivery report
#### this takes 0_non_fno_wl_yymmdd (created in 51) and generate 7-day delivery report
#### this takes sme yymmdd (created in 51) and generate 7-day delivery report
#### this also reads consolidated holdings file, create 0 my_holdings and generate 7-day delivery report


### this program reads the nse averages, bse averages, sme averages in spider/user/report folder 
### and combines them (after removing tags for trade to trade/ sm-st / z group etc)
### and writes to nse bse averages in input folder under sql stock market data

### nse bse averages is used by 53 multiple scrips to read averages and print them


### lenovo desktop file path
master_file_path = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Python Code from ChatGPT\\Master File\\CompanyMaster.csv"
directory = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Data Files\\"  # Replace with the path to your directory
output_path = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Python Code from ChatGPT\\Output File\\"
combined_file_path = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Delivery Price Combined Files\\"

column_order = ['NSE_Code','BSE_Code','Date','NSE_Delivery','BSE_Delivery']


#### Lenovo laptop file path


spider_path = "C:\\Spider\\User\\Report\\"

master_file_path = "C:\\Users\\sarda\\Dropbox\\SQL Stock Market Data\\Python Code from ChatGPT\\Master File\\CompanyMaster.csv"
directory = "C:\\Users\\sarda\\Dropbox\\SQL Stock Market Data\\Data Files\\"  # Replace with the path to your directory
output_path = "C:\\Users\\sarda\\Dropbox\\SQL Stock Market Data\\Python Code from ChatGPT\\Output File\\"
input_path = "C:\\Users\\sarda\\Dropbox\\SQL Stock Market Data\\Python Code from ChatGPT\\Input File\\"
combined_file_path = "C:\\Users\\sarda\\Dropbox\\SQL Stock Market Data\\Delivery Price Combined Files\\"

# Define the new column order
column_order = ["Exchange", "Code", "Name", "Close", "ema5", "ema10", "ema20", "ema50", "ema200", "75d high", "75d low"]

# Define the corresponding column types
column_types = {"Close": float, "ema5": float, "ema10": float, "ema20": float, "ema50": float, "ema200": float, 
                "75d high": float, "75d low": float}

columns_to_fill = ["Close", "ema5", "ema10", "ema20", "ema50", "ema200", "75d high", "75d low"]




nse_avg = spider_path + "NSE Averages.xls"
sme_avg = spider_path + "SME Averages.xls"
bse_avg = spider_path + "BSE Averages.xls"

input_file = input_path + "ShortListed Stocks.xlsx"

bse_sheet = "BSE Averages"
nse_sheet = "NSE Averages"

import os
import csv
import pandas as pd


nse_avg_df = pd.read_excel(nse_avg, names=column_order, dtype=column_types)
sme_avg_df = pd.read_excel(sme_avg, names=column_order, dtype=column_types)
bse_avg_df = pd.read_excel(bse_avg, names=column_order, dtype=column_types)
#input_file_df = pd.read_excel(input_file)


nse_avg_df[columns_to_fill] = nse_avg_df[columns_to_fill].fillna(0)
bse_avg_df[columns_to_fill] = bse_avg_df[columns_to_fill].fillna(0)
sme_avg_df[columns_to_fill] = sme_avg_df[columns_to_fill].fillna(0)

nse_sme_avg_df = pd.concat([nse_avg_df,sme_avg_df], ignore_index = True)

nse_avg_df = nse_sme_avg_df

#nse_avg_df.info()

nse_avg_df['Name'] = nse_avg_df['Name'].str.replace('_T$',"",regex=True)
nse_avg_df['Name'] = nse_avg_df['Name'].str.replace('_Z$',"",regex=True)
nse_avg_df['Name'] = nse_avg_df['Name'].str.replace(' SM$',"",regex=True)
nse_avg_df['Name'] = nse_avg_df['Name'].str.replace(' ST$',"",regex=True)


nse_avg_df['Name'] = nse_avg_df['Name'].str.strip()
nse_avg_df['Code'] = nse_avg_df['Code'].astype(str)

bse_avg_df['Code'] = bse_avg_df['Code'].astype(str)


# Assuming nse_avg_df and bse_avg_df are your DataFrames

# Replace 'NSE BSE Averages.xlsx' with the desired output file path
averages_file = input_path + 'NSE BSE Averages.xlsx'

# Use ExcelWriter to write to a new Excel file with two sheets
with pd.ExcelWriter(averages_file, engine='openpyxl', mode='w') as writer:
    nse_avg_df.to_excel(writer, sheet_name='NSE Averages', index=False)
    bse_avg_df.to_excel(writer, sheet_name='BSE Averages', index=False)

# Confirm the write operation
print(f"Data has been written to '{averages_file}' under the sheets 'NSE Averages' and 'BSE Averages'.")


# code_0.ipynb

import nbformat
from nbconvert.preprocessors import ExecutePreprocessor

# Load "56 Delivery comparision over 2 days.ipynb"
with open('56 Delivery comparision over 2 days.ipynb', 'r') as f:
    nb_content = nbformat.read(f, as_version=4)

# Execute the notebook
exec_preprocessor = ExecutePreprocessor(timeout=600, kernel_name='python3')
exec_preprocessor.preprocess(nb_content, {'metadata': {'path': './'}})

# Extract variables or functions from the executed notebook
executed_code = '\n'.join(cell['source'] for cell in nb_content.cells if cell.cell_type == 'code')

# Execute the extracted code in the current notebook
exec(executed_code)
