==> Python code for chatGpt/61 FnO BhavCopy for Trade Files.ipynb <==
import os
import csv
import pandas as pd

#directory_path = "C:\\Users\\User\\Dropbox\\SQL Stock Market Data\\Data Files\\"  # Replace with the path to your directory
#trade_file_path = "C:\\Users\\User\\Dropbox\\TS Personal\\Trade Files\\"

directory_path = "C:\\Users\\sarda\\Dropbox\\SQL Stock Market Data\\Data Files\\"  # Replace with the path to your directory
trade_file_path = "C:\\Users\\sarda\\Dropbox\\TS Personal\\Trade Files\\"

### file to save 

bhav_copy = trade_file_path + 'NSE FnO BhavCopy.xlsx'

### files to read  NSE FNO latest file from data files

nse_fno_file = sorted([file for file in os.listdir(directory_path) if file.startswith('NSE FNO')],reverse=True)[:1]
for file_name in nse_fno_file:
    file_path = os.path.join(directory_path, file_name)
    fno_df = pd.read_csv(file_path)

print('Processing ', nse_fno_file)


# Assuming fno_df is your DataFrame
fno_df['EXPIRY_DT'] = pd.to_datetime(fno_df['EXPIRY_DT'], format='%d-%b-%y')
reference_date = pd.to_datetime('1899-12-30')  # Choose a reference date

# Calculate the DaySerial column
fno_df['DaySerial'] = (fno_df['EXPIRY_DT'] - reference_date).dt.days

# Display the updated DataFrame


# Function to generate Code based on conditions
def generate_code(row):
    if row['OPTION_TYP'].lower() == 'xx':
        return f"{row['SYMBOL']}{int(row['DaySerial'])}XX0"
    else:
        return f"{row['SYMBOL']}{int(row['DaySerial'])}{row['OPTION_TYP'].upper()}{int(row['STRIKE_PR'])}"

# Apply the function to create the 'Code' column
fno_df['Code'] = fno_df.apply(generate_code, axis=1)

fno_df['close'] = fno_df["CLOSE"]

fno_df.to_excel(bhav_copy,index=False, engine='xlsxwriter')
print(f"DataFrame saved to {bhav_copy}")
